refactor: Issue #56 Phase 3 – CalendarScreen zerlegen, Duplikate entfernen, Type Safety
Phase 3a – CalendarScreen Komponenten-Extraktion:
- Neue Komponente CategoryTabBar.tsx für Kategorie-Tabs
- Neue Komponente TableHeader.tsx für sticky Header
- Neue Komponente PlantRowsContainer.tsx für Pflanzenzeilen
- CalendarScreen reduziert von 475 auf 129 Zeilen

Phase 3b – Duplikate bereinigen:
- CATEGORY_TABS zentral in src/constants/categoryTabs.ts
- CATEGORY_TABS_I18N (mit labelDe/labelEn) für i18n
- HALF_MONTH_NAMES zentral in monthHelper.ts statt duplikat in EditActivityModal
- Duplikate in CalendarScreen und AgendaScreen entfernt

Phase 3c – Type Safety:
- LanguageContext: t() return type → TranslationValue | string statt any
- Translations Record type: Record<Language, Translations>
- TranslationValue type für strings und string[]

Build: ✅ npm run build erfolgreich (921 kB bundle)
