+++ src/components/EditActivityModal.tsx
@@ -2,6 +2,7 @@ import React, { useState } from 'react';
 import { View, Text, StyleSheet, Modal, TouchableOpacity, TextInput, Alert } from 'react-native';
 import { useTheme } from '../hooks/useTheme';
 import { Activity } from '../types';
+import { HALF_MONTH_NAMES } from '../utils/monthHelper';
 
 interface EditActivityModalProps {
   visible: boolean;
@@ -108,13 +109,7 @@ export const EditActivityModal: React.FC<EditActivityModalProps> = ({
 };
 
 const getMonthName = (monthIndex: number): string => {
-  const months = [
-    'Jan 1-15', 'Jan 16-31', 'Feb 1-15', 'Feb 16-28', 'Mär 1-15', 'Mär 16-31',
-    'Apr 1-15', 'Apr 16-30', 'Mai 1-15', 'Mai 16-31', 'Jun 1-15', 'Jun 16-30',
-    'Jul 1-15', 'Jul 16-31', 'Aug 1-15', 'Aug 16-31', 'Sep 1-15', 'Sep 16-30',
-    'Okt 1-15', 'Okt 16-31', 'Nov 1-15', 'Nov 16-30', 'Dez 1-15', 'Dez 16-31',
-  ];
-  return months[monthIndex] || '';
+  return HALF_MONTH_NAMES[monthIndex] || '';
 };
 
 const styles = StyleSheet.create({
