+++ src/screens/CalendarScreen.tsx
@@ -1,22 +1,13 @@
 import React, { useState, useMemo, useRef } from 'react';
-import { View, Text, StyleSheet, ScrollView, ActivityIndicator, TouchableOpacity, Linking, Dimensions } from 'react-native';
+import { View, StyleSheet, Dimensions } from 'react-native';
 import { useTheme } from '../hooks/useTheme';
 import { usePlants } from '../contexts/PlantContext';
-import { PlantRow } from '../components/PlantRow';
 import { AddActivityModal } from '../components/AddActivityModal';
 import { EditActivityModal } from '../components/EditActivityModal';
-// import { AppHeader } from '../components/AppHeader'; // Temporär deaktiviert
-import { calculateActivityRows } from '../utils/activityLayout';
-import { PlantCategory } from '../types';
-
-type CategoryFilter = PlantCategory | 'all';
-
-const CATEGORY_TABS: { value: CategoryFilter; label: string; icon: string; color: string }[] = [
-  { value: 'all',       label: 'Alle',         icon: '🌿', color: '#4CAF50' },
-  { value: 'vegetable', label: 'Nutzpflanzen', icon: '🥦', color: '#F57C00' },
-  { value: 'flower',    label: 'Blumen',        icon: '🌸', color: '#E91E63' },
-  { value: 'tree',      label: 'Bäume',         icon: '🌳', color: '#2E7D32' },
-];
+import { CategoryTabBar } from '../components/CategoryTabBar';
+import { TableHeader } from '../components/TableHeader';
+import { PlantRowsContainer } from '../components/PlantRowsContainer';
+import { CategoryFilter } from '../constants/categoryTabs';
 
 export const CalendarScreen: React.FC = () => {
   const { theme } = useTheme();
@@ -30,11 +21,9 @@ export const CalendarScreen: React.FC = () => {
   const fixedScrollRef = useRef<ScrollView>(null);
   const headerScrollRef = useRef<ScrollView>(null);
 
-  // Responsive: Portrait zeigt 6 2-Monats-Slots, Landscape zeigt 24 Halbmonate
   const { width, height } = Dimensions.get('window');
   const isPortrait = height > width;
 
-  // Portrait: 6 Spalten (je 2 Monate), Landscape: 24 Halbmonate
   const months = useMemo(() => {
     if (isPortrait) {
       return ['Jan-Feb', 'Mär-Apr', 'Mai-Jun', 'Jul-Aug', 'Sep-Okt', 'Nov-Dez'];
@@ -47,12 +36,10 @@ export const CalendarScreen: React.FC = () => {
     ];
   }, [isPortrait]);
 
-  // Berechne aktuellen Halbmonat
   const now = new Date();
   const currentMonth = now.getMonth();
   const currentHalfMonth = currentMonth * 2 + (now.getDate() <= 15 ? 0 : 1);
 
-  // Sortiere und filtere Pflanzen nach Kategorie
   const sortedPlants = useMemo(() => {
     const filtered = activeCategory === 'all'
       ? plants
@@ -60,16 +47,8 @@ export const CalendarScreen: React.FC = () => {
     return filtered.sort((a, b) => a.name.localeCompare(b.name, 'de'));
   }, [plants, activeCategory]);
 
-  if (loading) {
-    return (
-      <View style={[styles.container, { backgroundColor: theme.background, justifyContent: 'center', alignItems: 'center' }]}> 
-        <ActivityIndicator size="large" color={theme.primary} />
-        <Text style={[styles.loadingText, { color: theme.text }]}>Lade Pflanzen...</Text>
-      </View>
-    );
-  }
   const selectedPlant = sortedPlants.find((p) => p.id === selectedPlantId) || null;
-  const selectedActivity = selectedPlant?.activities.find((a: any) => a.id === selectedActivityId) || null;
+  const selectedActivity = selectedPlant?.activities.find((a) => a.id === selectedActivityId) || null;
 
   const handleAddActivity = (type: string, startMonth: number, endMonth: number, color: string, label: string) => {
     if (selectedPlantId) {
@@ -79,7 +58,6 @@ export const CalendarScreen: React.FC = () => {
 
   const handlePressMonth = (plantId: string, monthIndex: number) => {
     setSelectedPlantId(plantId);
-    // Im Portrait-Modus: Konvertiere 2-Monats-Slot-Index zu Halbmonat
     setSelectedMonth(isPortrait ? monthIndex * 4 : monthIndex);
     setShowAddActivity(true);
   };
@@ -90,7 +68,7 @@ export const CalendarScreen: React.FC = () => {
     setShowEditActivity(true);
   };
 
-  const handleUpdateActivity = (activityId: string, updates: any) => {
+  const handleUpdateActivity = (activityId: string, updates: Record<string, unknown>) => {
     if (selectedPlantId) {
       updateActivity(selectedPlantId, activityId, updates);
     }
@@ -104,213 +82,27 @@ export const CalendarScreen: React.FC = () => {
 
   return (
     <View style={[styles.container, { backgroundColor: theme.background }]}>
-      {/* AppHeader temporär deaktiviert - Navigation ist jetzt im Haupt-Header */}
+      <CategoryTabBar activeCategory={activeCategory} onCategoryChange={setActiveCategory} />
 
-      {/* Kategorie-Tabs */}
-      <View style={[styles.tabBar, { borderBottomColor: theme.border, backgroundColor: theme.background }]}>
-        {CATEGORY_TABS.map((tab) => {
-          const isActive = activeCategory === tab.value;
-          return (
-            <TouchableOpacity
-              key={tab.value}
-              style={styles.tab}
-              onPress={() => setActiveCategory(tab.value)}
-            >
-              <View style={[
-                styles.iconBadge,
-                { backgroundColor: isActive ? tab.color : tab.color + '30' },
-              ]}>
-                <Text style={styles.tabIcon}>{tab.icon}</Text>
-              </View>
-              <Text style={[styles.tabLabel, { color: isActive ? tab.color : theme.textSecondary, fontWeight: isActive ? '700' : '400' }]}>
-                {tab.label}
-              </Text>
-            </TouchableOpacity>
-          );
-        })}
-      </View>
+      <TableHeader
+        months={months}
+        isPortrait={isPortrait}
+        currentHalfMonth={currentHalfMonth}
+        headerScrollRef={headerScrollRef}
+      />
 
-      <View style={styles.tableContainer}>
-        <View style={styles.fixedColumn}>
-          {/* Fixed Header - Pflanze */}
-          <View style={[styles.fixedHeaderCell, { borderColor: theme.border, backgroundColor: theme.surface }]}>
-            <Text style={[styles.headerText, { color: theme.text }]}>Pflanze</Text>
-          </View>
+      <PlantRowsContainer
+        sortedPlants={sortedPlants}
+        isPortrait={isPortrait}
+        currentHalfMonth={currentHalfMonth}
+        months={months}
+        onPressActivity={handlePressActivity}
+        onPressMonth={handlePressMonth}
+        fixedScrollRef={fixedScrollRef}
+        headerScrollRef={headerScrollRef}
+        loading={loading}
+      />
 
-          {/* Fixed Plant Names */}
-          <ScrollView
-            style={styles.fixedColumnScroll}
-            showsVerticalScrollIndicator={false}
-            scrollEnabled={false}
-            ref={(ref) => {
-              if (ref && !fixedScrollRef.current) {
-                fixedScrollRef.current = ref;
-              }
-            }}
-          >
-            {sortedPlants.length === 0 ? (
-              <View style={[styles.fixedPlantCell, { borderColor: theme.border }]}>
-                <Text style={[styles.emptyText, { color: theme.textSecondary }]}>-</Text>
-              </View>
-            ) : (
-              sortedPlants.map(plant => {
-                // Calculate same height as PlantRow
-                const activitiesWithRows = calculateActivityRows(plant.activities);
-                const maxRow = activitiesWithRows.reduce((max, a) => Math.max(max, a.row), 0);
-                const minHeight = Math.max(60, (maxRow + 1) * 28 + 8);
-                
-                return (
-                  <View key={plant.id} style={[styles.fixedPlantCell, { borderColor: theme.border, backgroundColor: theme.surface, minHeight }]}>
-                    <Text style={[styles.plantNameText, { color: theme.text }]} numberOfLines={2}>
-                      {plant.name}
-                    </Text>
-                  </View>
-                );
-              })
-            )}
-          </ScrollView>
-        </View>
-
-        <View style={styles.scrollableColumn}>
-          {/* Sticky Header */}
-          <ScrollView
-            horizontal
-            showsHorizontalScrollIndicator={false}
-            style={styles.stickyHeaderScroll}
-            scrollEnabled={false}
-            ref={headerScrollRef}
-          >
-            <View style={[styles.headerRow, { backgroundColor: theme.background }]}>
-              {months.map((month, index) => {
-                let isCurrentPeriod = false;
-
-                if (isPortrait) {
-                  // Im Portrait: Prüfe ob aktueller Halbmonat in diesem 2-Monats-Slot ist
-                  const slotStart = index * 4;
-                  const slotEnd = slotStart + 3;
-                  isCurrentPeriod = currentHalfMonth >= slotStart && currentHalfMonth <= slotEnd;
-                } else {
-                  // Im Landscape: Wie bisher
-                  isCurrentPeriod = index === currentHalfMonth;
-                }
-
-                return (
-                  <View
-                    key={index}
-                    style={[
-                      isPortrait ? styles.twoMonthCell : styles.monthCell,
-                      {
-                        borderColor: theme.border,
-                        backgroundColor: isCurrentPeriod ? theme.border : theme.surface
-                      }
-                    ]}
-                  >
-                    <Text style={[styles.monthText, { color: theme.textSecondary }]}>
-                      {isPortrait ? month : (index % 2 === 0 ? month : '')}
-                    </Text>
-                    {!isPortrait && (
-                      <Text style={[styles.halfMonthText, { color: theme.textSecondary }]}>
-                        {index % 2 === 0 ? '1' : '2'}
-                      </Text>
-                    )}
-                  </View>
-                );
-              })}
-              <View style={[styles.notesCell, { borderColor: theme.border, backgroundColor: theme.surface }]}>
-                <Text style={[styles.headerText, { color: theme.text }]}>Notizen</Text>
-              </View>
-            </View>
-          </ScrollView>
-
-          {/* Scrollable content */}
-          <ScrollView
-            horizontal
-            showsHorizontalScrollIndicator={true}
-            style={styles.horizontalScroll}
-            contentContainerStyle={styles.horizontalContent}
-            onScroll={(e) => {
-              if (headerScrollRef.current) {
-                headerScrollRef.current.scrollTo({ x: e.nativeEvent.contentOffset.x, animated: false });
-              }
-            }}
-            scrollEventThrottle={16}
-          >
-            <ScrollView
-              style={styles.verticalScroll}
-              showsVerticalScrollIndicator={true}
-              nestedScrollEnabled={true}
-              onScroll={(e) => {
-                if (fixedScrollRef.current) {
-                  fixedScrollRef.current.scrollTo({ y: e.nativeEvent.contentOffset.y, animated: false });
-                }
-              }}
-              scrollEventThrottle={16}
-            >
-              <View style={styles.tableWrapper}>
-              {/* Pflanzenzeilen mit Aktivitätsbalken */}
-            {sortedPlants.length === 0 ? (
-              <View style={styles.emptyState}>
-                <Text style={[styles.emptyText, { color: theme.textSecondary }]}>
-                  Noch keine Pflanzen vorhanden.
-                </Text>
-              </View>
-            ) : (
-              sortedPlants.map(plant => {
-                // Im Portrait: Konvertiere Halbmonate zu 2-Monats-Slots
-                const visibleActivities = plant.activities.map(activity => {
-                  if (isPortrait) {
-                    // Konvertiere Halbmonat-Index zu 2-Monats-Slot-Index
-                    // Slot 0: Halbmonate 0-3 (Jan-Feb)
-                    // Slot 1: Halbmonate 4-7 (Mär-Apr)
-                    // usw.
-                    const startSlot = Math.floor(activity.startMonth / 4);
-                    const endSlot = Math.floor(activity.endMonth / 4);
-                    return {
-                      ...activity,
-                      id: activity.id,
-                      startMonth: startSlot,
-                      endMonth: endSlot,
-                    };
-                  }
-                  // Landscape: Keine Änderung
-                  return activity;
-                });
-
-                const visiblePlant = {
-                  ...plant,
-                  activities: visibleActivities,
-                };
-
-                const handleActivityClick = (activityId: string) => {
-                  handlePressActivity(plant.id, activityId);
-                };
-
-                const handleMonthClick = (monthIndex: number) => {
-                  handlePressMonth(plant.id, monthIndex);
-                };
-
-                return (
-                  <PlantRow
-                    key={plant.id}
-                    plant={visiblePlant}
-                    onPressActivity={handleActivityClick}
-                    onPressMonth={handleMonthClick}
-                    onPressPlant={() => {}}
-                    totalMonths={months.length}
-                    currentHalfMonth={isPortrait ? Math.floor(currentHalfMonth / 4) : currentHalfMonth}
-                    monthOffset={0}
-                    cellWidth={isPortrait ? 60 : 40}
-                  />
-                );
-              })
-            )}
-            </View>
-          </ScrollView>
-        </ScrollView>
-        </View>
-      </View>
-
-      {/* Modals */}
       {selectedPlant && (
         <AddActivityModal
           visible={showAddActivity}
@@ -338,137 +130,5 @@ export const CalendarScreen: React.FC = () => {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-  },
-  tabBar: {
-    flexDirection: 'row',
-    borderBottomWidth: 1,
-    paddingHorizontal: 4,
-    paddingVertical: 8,
-  },
-  tab: {
-    flex: 1,
-    alignItems: 'center',
-    gap: 4,
-  },
-  iconBadge: {
-    width: 32,
-    height: 32,
-    borderRadius: 8,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  tabIcon: {
-    fontSize: 16,
-  },
-  tabLabel: {
-    fontSize: 9,
-  },
-  tableContainer: {
-    flex: 1,
-    flexDirection: 'row',
-  },
-  fixedColumn: {
-    width: 120,
-    zIndex: 10,
-  },
-  scrollableColumn: {
-    flex: 1,
-  },
-  stickyHeaderScroll: {
-    maxHeight: 60,
-  },
-  fixedHeaderCell: {
-    width: 120,
-    padding: 8,
-    borderWidth: 1,
-    justifyContent: 'center',
-    minHeight: 60,
-  },
-  fixedColumnScroll: {
-    flex: 1,
-  },
-  fixedPlantCell: {
-    width: 120,
-    minHeight: 60,
-    padding: 8,
-    borderWidth: 1,
-    borderTopWidth: 0,
-    justifyContent: 'center',
-  },
-  plantNameText: {
-    fontSize: 14,
-  },
-  horizontalScroll: {
-    flex: 1,
-  },
-  horizontalContent: {
-    flexGrow: 1,
-  },
-  verticalScroll: {
-    flex: 1,
-  },
-  tableWrapper: {
-    minWidth: '100%',
-  },
-  navButtons: {
-    flexDirection: 'row',
-    gap: 8,
-  },
-  navButton: {
-    padding: 8,
-  },
-  navButtonText: {
-    fontSize: 20,
-    fontWeight: 'bold',
-  },
-  loadingText: {
-    marginTop: 12,
-    fontSize: 16,
-  },
-  headerRow: {
-    flexDirection: 'row',
-  },
-  monthCell: {
-    width: 40,
-    padding: 4,
-    borderWidth: 1,
-    borderLeftWidth: 0,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  twoMonthCell: {
-    width: 60,
-    padding: 4,
-    borderWidth: 1,
-    borderLeftWidth: 0,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  notesCell: {
-    width: 120,
-    padding: 8,
-    borderWidth: 1,
-    borderLeftWidth: 0,
-    justifyContent: 'center',
-  },
-  headerText: {
-    fontSize: 14,
-    fontWeight: '600',
-  },
-  monthText: {
-    fontSize: 10,
-    fontWeight: '500',
-  },
-  halfMonthText: {
-    fontSize: 8,
-  },
-  emptyState: {
-    padding: 40,
-    alignItems: 'center',
-  },
-  emptyText: {
-    fontSize: 14,
-    textAlign: 'center',
-    marginBottom: 4,
   },
 });
